pyproject.toml
[tool.poetry]
name = "ale-test-task"
version = "0.1.0"
description = ""
authors = ["Vlad Maslianko <masliankovladik@gmail.ocm>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "^0.115.3"
requests = "^2.32.3"
uvicorn = "^0.32.0"
openai = "^1.52.2"
python-dotenv = "^1.0.1"
redis = "^5.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
pytest-asyncio = "^0.24.0"
fastapi = "^0.115.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
